Utility class to expose a collection of Ref objects as RealPositionable,
even if they are not. The user can specify the coordinates with a list
of functions, and the resulting wrapper collection can be used in
KD-tree searches.

The KdTreeWrapper class does the job without
exposing the wrapper classes. But right not it only exposes the
function to perform searches inside a convex polytope.